Main: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@47924 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/xml/ovfreader.cpp
@@ -281,11 +281,11 @@ void OVFReader::HandleDiskSection(const xml::ElementNode *pReferencesElem,
         // suggest a size in megabytes to help callers with progress reports
         d.ulSuggestedSizeMB = 0;
         if (d.iCapacity != -1)
-            d.ulSuggestedSizeMB = d.iCapacity / _1M;
+            d.ulSuggestedSizeMB = (uint32_t)(d.iCapacity / _1M);
         else if (d.iPopulatedSize != -1)
-            d.ulSuggestedSizeMB = d.iPopulatedSize / _1M;
+            d.ulSuggestedSizeMB = (uint32_t)(d.iPopulatedSize / _1M);
         else if (d.iSize != -1)
-            d.ulSuggestedSizeMB = d.iSize / _1M;
+            d.ulSuggestedSizeMB = (uint32_t)(d.iSize / _1M);
         if (d.ulSuggestedSizeMB == 0)
             d.ulSuggestedSizeMB = 10000;         // assume 10 GB, this is for the progress bar only anyway
